Split config and event handling
