+++ pipeline/src/pipeline-config/pipelineDatabase.ts
@@ -1,57 +1,10 @@
 import { PostgresClient } from '@jvalue/node-dry-pg'
-import { ClientBase, PoolConfig, QueryResult } from 'pg'
+import { PoolConfig } from 'pg'
 
-import { POSTGRES_DB, POSTGRES_HOST, POSTGRES_PORT, POSTGRES_PW, POSTGRES_SCHEMA, POSTGRES_USER } from '../env'
-import { PipelineConfig, PipelineConfigDTO } from './model/pipelineConfig'
+import { POSTGRES_DB, POSTGRES_HOST, POSTGRES_PORT, POSTGRES_PW, POSTGRES_USER } from '../env'
 
-const CONFIG_TABLE_NAME = 'PipelineConfigs'
-const OUTBOX_TABLE_NAME = 'outbox'
-
-const CREATE_CONFIG_TABLE_STATEMENT = `
-  CREATE TABLE IF NOT EXISTS "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}" (
-  "id" bigint NOT NULL GENERATED ALWAYS AS IDENTITY,
-  "datasourceId" bigint NOT NULL,
-  "func" varchar NOT NULL,
-  "author" varchar,
-  "displayName" varchar NOT NULL,
-  "license" varchar,
-  "description" varchar,
-  "createdAt" timestamp,
-  CONSTRAINT "Data_pk_${POSTGRES_SCHEMA}_${CONFIG_TABLE_NAME}" PRIMARY KEY (id)
-)`
-const INSERT_CONFIG_STATEMENT = `
-  INSERT INTO "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}"
-  ("datasourceId", "func", "author", "displayName", "license", "description", "createdAt")
-  VALUES ($1, $2, $3, $4, $5, $6, $7)
-  RETURNING *`
-const UPDATE_CONFIG_STATEMENT = `
-  UPDATE "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}"
-  SET "datasourceId"=$2, "func"=$3, "author"=$4, "displayName"=$5, "license"=$6, "description"=$7
-  WHERE id=$1`
-const GET_CONFIG_STATEMENT = `
-  SELECT * FROM "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}" WHERE "id" = $1`
-const GET_ALL_CONFIGS_STATEMENT = `
-  SELECT * FROM "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}"`
-const GET_ALL_CONFIGS_BY_DATASOURCE_ID_STATEMENT = `
-  SELECT * FROM "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}" WHERE "datasourceId" = $1`
-const DELETE_CONFIG_STATEMENT = `
-  DELETE FROM "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}" WHERE "id" = $1 RETURNING *`
-const DELETE_ALL_CONFIGS_STATEMENT = `
-  DELETE FROM "${POSTGRES_SCHEMA}"."${CONFIG_TABLE_NAME}" RETURNING *`
-
-const CREATE_OUTBOX_TABLE_STATEMENT = `
-CREATE TABLE "${POSTGRES_SCHEMA}"."${OUTBOX_TABLE_NAME}"(
-  id uuid NOT NULL CONSTRAINT "Data_pk_${POSTGRES_SCHEMA}_${OUTBOX_TABLE_NAME}" PRIMARY KEY DEFAULT gen_random_uuid(),
-  routing_key varchar(255) NOT NULL,
-  payload jsonb NOT NULL
-);
-`
-const INSERT_EVENT_STATEMENT = `
-INSERT INTO "${POSTGRES_SCHEMA}"."${OUTBOX_TABLE_NAME}"
-  ("routing_key", "payload")
-  VALUES ($1, $2)
-  RETURNING id
-`
+import { createPipelineConfigTable } from './pipelineConfigRepository'
+import { createPipelineEventTable } from './pipelineEventRepository'
 
 const POOL_CONFIG: PoolConfig = {
   host: POSTGRES_HOST,
@@ -64,118 +17,12 @@ const POOL_CONFIG: PoolConfig = {
   connectionTimeoutMillis: 2000
 }
 
-interface DatabasePipeline {
-  id: string
-  datasourceId: string
-  func: string
-  author: string
-  displayName: string
-  license: string
-  description: string
-  createdAt: Date
-}
-
 export async function init (retries: number, backoffMs: number): Promise<PostgresClient> {
   const postgresClient = new PostgresClient(POOL_CONFIG)
   await postgresClient.waitForConnection(retries, backoffMs)
   await postgresClient.transaction(async client => {
-    await client.query(CREATE_CONFIG_TABLE_STATEMENT)
-    await client.query(CREATE_OUTBOX_TABLE_STATEMENT)
+    await createPipelineConfigTable(client)
+    await createPipelineEventTable(client)
   })
   return postgresClient
-}
-
-export async function create (client: ClientBase, config: PipelineConfigDTO): Promise<PipelineConfig> {
-  const values = [
-    config.datasourceId,
-    config.transformation.func,
-    config.metadata.author,
-    config.metadata.displayName,
-    config.metadata.license,
-    config.metadata.description,
-    new Date()
-  ]
-  const { rows } = await client.query(INSERT_CONFIG_STATEMENT, values)
-  return toPipelineConfig(rows[0])
-}
-
-export async function get (client: ClientBase, id: number): Promise<PipelineConfig | undefined> {
-  const resultSet = await client.query(GET_CONFIG_STATEMENT, [id])
-  if (resultSet.rowCount === 0) {
-    return undefined
-  }
-  const content = toPipelineConfigs(resultSet)
-  return content[0]
-}
-
-export async function getAll (client: ClientBase): Promise<PipelineConfig[]> {
-  const resultSet = await client.query(GET_ALL_CONFIGS_STATEMENT, [])
-  return toPipelineConfigs(resultSet)
-}
-
-export async function getByDatasourceId (client: ClientBase, datasourceId: number): Promise<PipelineConfig[]> {
-  const resultSet = await client.query(GET_ALL_CONFIGS_BY_DATASOURCE_ID_STATEMENT, [datasourceId])
-  return toPipelineConfigs(resultSet)
-}
-
-export async function update (client: ClientBase, id: number, config: PipelineConfigDTO): Promise<void> {
-  const values = [
-    id,
-    config.datasourceId,
-    config.transformation.func,
-    config.metadata.author,
-    config.metadata.displayName,
-    config.metadata.license,
-    config.metadata.description
-  ]
-  const result = await client.query(UPDATE_CONFIG_STATEMENT, values)
-  if (result.rowCount === 0) {
-    throw new Error(`Could not find config with ${id} to update`)
-  }
-}
-
-export async function deleteById (client: ClientBase, id: number): Promise<PipelineConfig> {
-  const result = await client.query(DELETE_CONFIG_STATEMENT, [id])
-  if (result.rowCount === 0) {
-    throw new Error(`Could not find config with ${id} to delete`)
-  }
-  const content = toPipelineConfigs(result)
-  return content[0]
-}
-
-export async function deleteAll (client: ClientBase): Promise<PipelineConfig[]> {
-  const result = await client.query(DELETE_ALL_CONFIGS_STATEMENT, [])
-  return toPipelineConfigs(result)
-}
-
-function toPipelineConfig (dbResult: DatabasePipeline): PipelineConfig {
-  return {
-    id: +dbResult.id,
-    datasourceId: +dbResult.datasourceId,
-    transformation: {
-      func: dbResult.func
-    },
-    metadata: {
-      author: dbResult.author,
-      displayName: dbResult.displayName,
-      license: dbResult.license,
-      description: dbResult.description,
-      creationTimestamp: dbResult.createdAt
-    }
-  }
-}
-
-function toPipelineConfigs (resultSet: QueryResult<DatabasePipeline>): PipelineConfig[] {
-  const configs: PipelineConfig[] = []
-  for (const row of resultSet.rows) {
-    const config = toPipelineConfig(row)
-    configs.push(config)
-  }
-  return configs
-}
-
-export async function insertEvent (client: ClientBase, routingKey: string, payload: unknown): Promise<string> {
-  const { rows } = await client.query(INSERT_EVENT_STATEMENT, [routingKey, payload])
-  console.log(typeof rows[0].id)
-  return rows[0].id
 }
